# exchange-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Exchange Rate Application",
    "description": "This is an exchange rate application where users can sign up for an account, add transactions, see their previous transactions, and get the current exchange rate calculated according to the past 3 days. Users can also check the conversion rate and amount for buying or selling USD and LBP.",
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000/"
    }
  ],
...
  "paths": {
...
    "/predictions": {
      "get": {
        "tags": [
          "Predictions"
        ],
        "summary": "Get the predicted exchange rate for a future date",
        "operationId": "predictRate",
        "parameters": [
          {
            "name": "YEAR",
            "in": "query",
            "required": true,
            "description": "End year, if null, today's date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 2024
            }
          },
          {
            "name": "MONTH",
            "in": "query",
            "required": true,
            "description": "End month, if null, today's date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 4
            }
          },
          {
            "name": "DAY",
            "in": "query",
            "required": true,
            "description": "End day, if null, today's date is used.",
            "schema": {
              "type": "integer",
              "nullable": true,
              "example": 25
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "rate_prediction": {
                      "type": "number",
                      "format": "number",
                      "example": 89700.5
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters provided"
          },
          "403": {
            "description": "Cannot predict for previous days. Predict into the future only"
          }
        }
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "JWTToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "JSON Web Token (JWT) allow users to link transactions to themselves.\nAdd the JWT token in the format `Bearer <token>`.\n"
      },
      "None": {
        "type": "http",
        "scheme": "none",
        "description": "Authentication is not needed always.\n"
      }
    }
  }
}